Add 2 seconds delay
Sometimes the service fails to set new values on boot. No error is reported, the driver seems to accept the new values, but testing shows that the new values are not effectively in use. This is likely a driver bug. Adding a delay helps make sure the driver is ready to set new OverDrive tables. This is a hack, 2 seconds have proved enough in testing on a laptop.