Fix resume from suspend
The kernel driver resets OverDrive/PowerPlay settings on suspend, so reapplying settings is needed on resume. A systemd timer was added to better control scheduling to apply the settings both at boot time and when resuming from sleep. A big change to how the service is executed was needed to accommodate the new scheduling.