Add timer to reapply settings daily at 6:00
When the kernel driver crashes, the OverDrive/PowerPlay tables are reset to default values. Driver crashes are common and happen often, which means long-running systems will be reset to defaults most of the time. Reapplying our settings daily helps alleviate the problem on systems that don't reboot/suspend for weeks.